POC using registered GUI's for the return function.

+++ Plugin/src/main/java/net/bteuk/network/gui/plotsystem/PlotMenu.java
@@ -1,6 +1,5 @@
 package net.bteuk.network.gui.plotsystem;
 
-import net.bteuk.network.gui.BuildGui;
 import net.bteuk.network.gui.GuiProvider;
 import net.bteuk.network.gui.NetworkRefreshableGui;
 import net.bteuk.network.sql.PlotSQL;
@@ -9,9 +8,12 @@ import net.bteuk.network.utils.Utils;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextDecoration;
+import org.btuk.minecraft.gui.GuiFactory;
+import org.btuk.minecraft.gui.GuiOpenContext;
 import org.bukkit.Material;
 
 import java.util.ArrayList;
+import java.util.Collections;
 
 public class PlotMenu extends NetworkRefreshableGui {
 
@@ -101,12 +103,13 @@ public class PlotMenu extends NetworkRefreshableGui {
                         Utils.title("Return"),
                         Utils.line("Open the building menu.")),
                 (NetworkUser u) -> {
-                    // Delete this gui.
-                    this.delete();
-
-                    // Switch to plot info.
-                    u.mainGui = new BuildGui(provider, u);
-                    u.mainGui.open(u.player);
+                    GuiFactory guiFactory = getManager().getReturnGui(PlotSystemGuiType.PLOT_MENU);
+                    if (guiFactory != null) {
+                        // Switch back to the build menu.
+                        this.delete();
+                        u.mainGui = guiFactory.create(new GuiOpenContext(u.player, Collections.emptyMap()));
+                        u.mainGui.open(u.player);
+                    }
                 });
     }
