+++ Plugin/src/main/java/net/bteuk/network/gui/plotsystem/PlotsystemLocations.java
@@ -1,7 +1,6 @@
 package net.bteuk.network.gui.plotsystem;
 
 import net.bteuk.network.api.entity.NetworkLocation;
-import net.bteuk.network.gui.BuildGui;
 import net.bteuk.network.gui.GuiProvider;
 import net.bteuk.network.gui.NetworkRefreshableGui;
 import net.bteuk.network.papercore.LocationAdapter;
@@ -13,11 +12,14 @@ import net.bteuk.network.utils.Utils;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextDecoration;
+import org.btuk.minecraft.gui.GuiFactory;
+import org.btuk.minecraft.gui.GuiOpenContext;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.World;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Objects;
 
 public class PlotsystemLocations extends NetworkRefreshableGui {
@@ -117,13 +119,13 @@ public class PlotsystemLocations extends NetworkRefreshableGui {
                         Utils.title("Return"),
                         Utils.line("Open the building menu.")),
                 (NetworkUser u) -> {
-
-                    // Delete this gui.
-                    this.delete();
-
-                    // Switch to plot info.
-                    u.mainGui = new BuildGui(provider, u);
-                    u.mainGui.open(u.player);
+                    GuiFactory guiFactory = getManager().getReturnGui(PlotSystemGuiType.PLOT_SYSTEM_LOCATIONS);
+                    if (guiFactory != null) {
+                        // Switch back to the build menu.
+                        this.delete();
+                        u.mainGui = guiFactory.create(new GuiOpenContext(u.player, Collections.emptyMap()));
+                        u.mainGui.open(u.player);
+                    }
                 });
     }
